feat : socket connect 로직 분리
